float64 to int64 + bug fix proxy_id comparison

+++ configUtils/ParseDynamicConfig.go
@@ -18,7 +18,7 @@ func InjectMappingRules(fullConfigBytes []byte, fixedConfig model.Configuration)
 			proxy_config := res.(map[string]interface{})["proxy_config"].(map[string]interface{})
 			content := proxy_config["content"].(map[string]interface{})
 			proxy := content["proxy"].(map[string]interface{})
-			id := proxy["id"].(float64)
+			id := proxy["id"].(int64)
 			fixedProxy := fixedConfig.ProxyConfigsOuter[index].ProxyConfig.Content.Proxy
 			if fixedProxy.Id != id {
 				panic("WhaAAAAAAAAAAAAAAAAA " + fmt.Sprint(id) + " != " + fmt.Sprint(fixedProxy.Id))
@@ -36,7 +36,7 @@ func InjectMappingRules(fullConfigBytes []byte, fixedConfig model.Configuration)
 		services := config["services"]
 		for index, res := range services.([]interface{}) {
 			proxy := res.(map[string]interface{})["proxy"].(map[string]interface{})
-			id := proxy["id"].(float64)
+			id := proxy["id"].(int64)
 			fixedProxy := fixedConfigCB.Config.Services[index].Proxy
 			if fixedProxy.Id != id {
 				panic("WhaAAAAAAAAAAAAAAAAA " + fmt.Sprint(id) + " != " + fmt.Sprint(fixedProxy.Id))
@@ -53,7 +53,7 @@ func InjectMappingRules(fullConfigBytes []byte, fixedConfig model.Configuration)
 		services := fullConfig.FullConfigContainer["services"].([]interface{})
 		for index, res := range services {
 			proxy := res.(map[string]interface{})["proxy"].(map[string]interface{})
-			id := proxy["id"].(float64)
+			id := proxy["id"].(int64)
 			fixedProxy := fixedConfigCC.Services[index].Proxy
 			if fixedProxy.Id != id {
 				panic("WhaAAAAAAAAAAAAAAAAA " + fmt.Sprint(id) + " != " + fmt.Sprint(fixedProxy.Id))
@@ -66,7 +66,7 @@ func InjectMappingRules(fullConfigBytes []byte, fixedConfig model.Configuration)
 		fixedConfigSS = fixedConfig.ProxyConfigsOuter[0].ProxyConfig.Content
 		//==========================================
 		fixedProxy := fixedConfigSS.Proxy
-		id := fullConfig.FullConfigContainer["proxy"].(map[string]interface{})["id"].(float64)
+		id := fullConfig.FullConfigContainer["proxy"].(map[string]interface{})["id"].(int64)
 		if fixedProxy.Id != id {
 			panic("WhaAAAAAAAAAAAAAAAAA " + fmt.Sprint(id) + " != " + fmt.Sprint(fixedProxy.Id))
 		}
